fix loading page
