+++ styles/loadMixin.styl
@@ -4,7 +4,7 @@ load-start() {
 }
 
 load-end($delayTime) {
-  transition:all .25s;
+  transition all .25s
   transform translateY(0)
   transition-delay $delayTime
   opacity 1
